Change class ShellSelectionWrapper into record ChartsRenderingDataWrapper
